\name{add_means_and_errors}
\alias{add_means_and_errors}
\title{Add means and standard errors to de_data by condition}
\usage{
add_means_and_errors(de_data, conditions)
}
\arguments{
  \item{de_data}{list containing result data frame, DE test
  output and column indices}

  \item{conditions}{vector of condition names}
}
\value{
list, de_data with added mean and error columns for each
condition
}
\description{
Add means and standard errors to de_data by condition
}

